requirements.txt
flask
pandas
google-generativeai
reportlab
openpyxl
genai
gunicorn